Proect calls/requirements.txt
Django==2.1.3
django-crispy-forms==1.7.2
django-filter==2.1.0
Pillow==5.3.0
psycopg2==2.8.3
pytz==2018.4
reportlab==3.5.23
djangorestframework==3.11.2